docs: add startLTS usage examples
I reach to `node-core-utils` for automating the LTS release commit but I
couldn't get it to work due to the lack of a proper usage example.

I'm adding examples to both the markdown docs and the cli help output which are
the places I would have loved to have found these.

Signed-off-by: Ruy Adorno <ruyadorno@google.com>